- disk.cpp looks for HFS partition info in the disk image; this makes it   possible to, for example, use MacOS-partitioned hard disks and removable   media under B2/Unix even if the OS doesn't understand Mac partition maps   by specifying the appropriate block device name as a Mac volume - fixed typo in audio_dummy.cpp - added minimally required UDP tunneling code to ether_dummy.cpp - main_unix.cpp: if pthreads are not supported, we trigger the Ethernet   interrupt in the 60Hz ticker; this makes UDP tunneling work under   NetBSD/m68k (as the only form of networking)

+++ BasiliskII/src/cdrom.cpp
@@ -132,7 +132,7 @@ struct cdrom_drive_info {
 	void *fh;			// File handle
 	int block_size;		// CD-ROM block size
 	int twok_offset;	// Offset of beginning of 2K block to last Prime position
-	uint32 start_byte;	// Start of HFS partition on disk
+	loff_t start_byte;	// Start of HFS partition on disk
 	bool to_be_mounted;	// Flag: drive must be mounted in accRun
 	bool mount_non_hfs;	// Flag: Issue disk-inserted events for non-HFS disks
